+++ packages/_vue3-migration-test/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.0-alpha.16](https://github.com/empathyco/x/compare/vue3-migration-test@1.0.0-alpha.15...vue3-migration-test@1.0.0-alpha.16) (2024-05-28)
+
+
+### Features
+
+* **base-modal:** migrate base-modal component and side effects (#1479) ([610ec16](https://github.com/empathyco/x/commit/610ec16e20b10a344936b2914c61085e22a09dfd))
+* **sliding-panel:** migrate sliding-panel component to composition API (#1485) ([bf9e1a3](https://github.com/empathyco/x/commit/bf9e1a38364a23402562a19ce8095452da5f02d7))
+
+
+
+## [1.0.0-alpha.15](https://github.com/empathyco/x/compare/vue3-migration-test@1.0.0-alpha.14...vue3-migration-test@1.0.0-alpha.15) (2024-05-28)
+
+
+### Code Refactoring
+
+* migrate `BaseVariableColumnGrid` to Composition API (#1482) ([b7e292a](https://github.com/empathyco/x/commit/b7e292ae55560f557dd7db3cfbceed993c7b318e))
+
+
+
 ## [1.0.0-alpha.14](https://github.com/empathyco/x/compare/vue3-migration-test@1.0.0-alpha.13...vue3-migration-test@1.0.0-alpha.14) (2024-05-28)